+++ pybillomat/client_properties.py
@@ -12,6 +12,7 @@ import xml.etree.ElementTree as ET
 import errors
 from bunch import Bunch
 from http import Url
+from _itemsiterator import ItemsIterator
 
 
 class ClientProperty(Bunch):
@@ -276,7 +277,7 @@ class ClientProperties(list):
             )
 
 
-class ClientPropertiesIterator(object):
+class ClientPropertiesIterator(ItemsIterator):
     """
     Iterates over all found properties
     """
@@ -287,7 +288,7 @@ class ClientPropertiesIterator(object):
         """
 
         self.conn = conn
-        self.client_properties = ClientProperties(self.conn)
+        self.items = ClientProperties(self.conn)
         self.per_page = per_page
         self.search_params = Bunch(
             client_id = None,
@@ -317,7 +318,7 @@ class ClientPropertiesIterator(object):
 
     def load_page(self, page):
 
-        self.client_properties.search(
+        self.items.search(
             client_id = self.search_params.client_id,
             client_property_id = self.search_params.client_property_id,
             order_by = self.search_params.order_by,
@@ -328,69 +329,5 @@ class ClientPropertiesIterator(object):
             page = page,
             per_page = self.per_page
         )
-
-
-    def __len__(self):
-        """
-        Returns the count of found properties
-        """
-
-        return self.client_properties.total or 0
-
-
-    def __iter__(self):
-        """
-        Iterate over all found items
-        """
-
-        if not self.client_properties.pages:
-            return
-
-        for page in range(1, self.client_properties.pages + 1):
-            if not self.client_properties.page == page:
-                self.load_page(page = page)
-            for client in self.client_properties:
-                yield client
-
-
-    def __getitem__(self, key):
-        """
-        Returns the requested property from the pool of found properties
-        """
-
-        # List-Ids
-        all_list_ids = range(len(self))
-        requested_list_ids = all_list_ids[key]
-        is_list = isinstance(requested_list_ids, list)
-        if not is_list:
-            requested_list_ids = [requested_list_ids]
-        assert isinstance(requested_list_ids, list)
-
-        result = []
-
-        for list_id in requested_list_ids:
-
-            # In welcher Seite befindet sich die gewünschte ID?
-            for page_nr in range(1, self.client_properties.pages + 1):
-                max_list_id = (page_nr * self.client_properties.per_page) - 1
-                if list_id <= max_list_id:
-                    page = page_nr
-                    break
-            else:
-                raise AssertionError()
-
-            # Load page if neccessary
-            if not self.client_properties.page == page:
-                self.load_page(page = page)
-
-            # Add equested client-property to result
-            list_id_in_page = list_id - ((page - 1) * self.client_properties.per_page)
-            result.append(self.client_properties[list_id_in_page])
-
-        if is_list:
-            return result
-        else:
-            return result[0]
-